# Film Manager Implementation/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  version: 1.0.0
  description: REST APIs that alllows to interact with the Film Manager Service
...
paths:
...
  "/api/films/public":
    get:
      tags: 
      - "api/films/public"
      operationId: getPublicFilms
      summary: retrieve a list of public films.
      description: A generic user can retrieve a list of public films added by other users. 
        A pagination machanism is used in order to limit the size of the messages.
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example: 
                {
                  "totalPages": 3,
                  "currentPage": 1,
                  "films": [
                      {
                          "id": 2,
                          "title": "Heaven's Feel",
                          "owner": 2,
                          "private": false,
                          "reviews": "/api/films/public/2/reviews",
                          "self": "/api/films/public/2"
                      },
                      {
                          "id": 3,
                          "title": "You Can (Not) Redo",
                          "owner": 1,
                          "private": false,
                          "reviews": "/api/films/public/3/reviews",
                          "self": "/api/films/public/3"
                      }
                  ],
                  "next": "/api/films/public?pageNo=2"
              }
        "500":
          description: Internal Server Error
...
  "/api/films/public/{filmId}/reviews/{reviewerId}":
    get:
      tags: 
      - "api/films/public/{filmId}/reviews/{reviewerId}"
      operationId: getSingleReview
      summary: retrieve a specific review of a public film.
      description: A generic user can retrieve a specific completed review of a public film.
        An authenticated user can retrive an incomplete review if it's assigned to him 
        or if he is the owner of the film
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
        - name: reviewerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Review"
              examples:
                example1:
                  value:
                    filmId: 2
                    reviewerId: 5
                    completed: true
                    reviewDate: "2022-03-04"
                    rating": 10
                    review: This film is a perfect conclusion for the trilogy.
                    self: "/api/films/public/2/reviews/5"
              
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    put:
      tags: 
      - "api/films/public/{filmId}/reviews/{reviewerId}"
      security:
      - cookieAuth: []
      operationId: completeReview
      summary: Complete a review
      description: This operation allows to set the review, identified by the pair filmId, reviewerId, as completed.
         The completed property is set to true and the reviewDate, rating and review properties are changed
         according to the values specified by the user that completed the review.
      parameters:
      - name: filmId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example: 4
      - name: reviewerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      requestBody:
        description: The completed review 
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Review"
            example: 
              {
                "filmId": 4,
                "reviewerId": 1,
                "completed": true,
                "reviewDate": "2025-01-04",
                "rating": 9,
                "review": "Amazing story!",
              }

                
        required: true
            
      responses:
        "204":
          description: Succesfully completed
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    delete:
      tags: 
      - "api/films/public/{filmId}/reviews/{reviewerId}"
      security:
      - cookieAuth: []
      operationId: deleteReviewInvitation
      summary: Delete a review invitation
      description: This operation can be performed only by the film's owner on a review that is not yet completed.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 6
        - name: reviewerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "204":
          description: Successfully Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict (e.g. the review is already completed)
        "500":
          description: Internal Server Error
...
  "/api/users":
    get:
      tags: 
      - "api/users"
      security:
      - cookieAuth: []
      summary: Get informations about users
      description: An authenticated user can retrieve informations about all the other users (Password excluded)
      operationId: getUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
              example: 
                [
                  {
                      "id": 1,
                      "name": "User",
                      "email": "user.dsp@polito.it",
                      "self": "/api/users/1"
                  },
                  {
                      "id": 2,
                      "name": "Frank",
                      "email": "frank.stein@polito.it",
                      "self": "/api/users/2"
                  },
                  {
                      "id": 3,
                      "name": "Karen",
                      "email": "karen.makise@polito.it",
                      "self": "/api/users/3"
                  },
                  {
                      "id": 4,
                      "name": "Rene",
                      "email": "rene.regeay@polito.it",
                      "self": "/api/users/4"
                  },
                  {
                      "id": 5,
                      "name": "Beatrice",
                      "email": "beatrice.golden@polito.it",
                      "self": "/api/users/5"
                  },
                  {
                      "id": 6,
                      "name": "Arthur",
                      "email": "arthur.pendragon@polito.it",
                      "self": "/api/users/6"
                  }
                  ]
        '401':
          description: Unauthorized
        '404':
          description: Not Found
  "/api/users/{userId}":
    get:
      tags: 
      - "api/users/{userId}"
      security:
      - cookieAuth: []
      summary: Get informations about a specific user 
      description: An authenticated user can retrieve informations about a user idetified by userId (Password excluded)
      operationId: getSingleUser
      parameters:
      - name: userId
        in: path
        description: unique identifier of the user to get
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
              example: 
                {
                    "id": 1,
                    "name": "User",
                    "email": "user.dsp@polito.it",
                    "self": "/api/users/1"
                }
        '401':
          description: Unauthorized
        '404':
          description: Not Found
  "/api/users/authenticator":
    post:
      tags:
      - "api/users/authenticator"
      summary: Logs a user in
      description: The user who wants to log in sends the user data to the
        authenticator which performs the operation. If the request for the login of a new user
        comes from an already authenticated user, the previous user is first logged out.
      operationId: authenticateUser
      requestBody:
        description: The data of the user who wants to perform log in. The data
          structure must contain email and password.
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/User"
            example: 
              {
                "email":"user.dsp@polito.it",
                "password":"password"
            }
        required: true
      responses:
        '200':
          description: Succesfully Authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
                example: token=scal110; Path=/; HttpOnly
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
              example: 
                {
                    "id": 1,
                    "name": "User",
                    "email": "user.dsp@polito.it",
                    "self": "/api/users/1"
                }
        '401':
          description: Unauthorized    
...
components:
  schemas:
...
    User:
      description: schema for User data structure
      type: object
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON schema used to validate the user data structure
        id:
          type: integer
          description: The unique identifier of the user
        name:
          type: string
          description: the username of the user
        email:
          type: string
          format: email
          description: the user's email.
        password:
          type: string
          minLength: 6
          maxLength: 20
          description: the user's password.
        self:
          type: string
          format: uri
          description:  URI to access user-specific details.

      required:
        - email
        - password
    Users:
      type: array
      items:
        "$ref": "#/components/schemas/User"
      description: Array of User objects representing the list of users.
    Film:
      type: object
      description: The schema for the film data structure
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: The path to the schema that describes the film data structure
        id:
          type: integer
          description: The film's unique identifier
        title:
          type: string
          description: The film's title
        owner:
          type: integer
          description: The id of the user that created the film data structure
        private:
          type: boolean
          default: true
          description: It's true if the owner is the only one that can access the film, false otherwise
        watchDate:
          type: string
          format: date
          description: The date when the owner watched the Film
        rating:
          type: integer
          minimum: 0
          maximum: 10
          description: the rating that the owner has given to the film
        favourite:
          type: boolean
          default: false
          description: It's true if the film is among the owner favourites, false otherwise
        reviews:
          type: string
          format: uri
          description: the reviews that other users have given to the film
        self:
          type: string
          format: uri
          description: URI to access the specific film resource.
      required:
        - title
        - owner
        - private
    Films:
      type: object
      additionalProperties: false
      description: the schema for the films data structure
      properties:
        "$schema":
          type: string
          description: The path to the schema used to validate the films data structure
        totalPages:
          type: integer
          minimum: 1
          description: the total number of pages
        currentPage:
          type: integer
          minimum: 1
          description: the number of the currentPage
        films:
          type: array
          items:
            "$ref": "#/components/schemas/Film"
          description: the list of films on the current page
        next:
          type: string
          format: uri
          description: URI for accessing the next page
      required:
        - totalPages
    Review:
      type: object
      description: The schema for the review data structure
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: the path to the scham used to validate the review data structure
        filmId:
          type: integer
          description: unique identifier of the film for which a review invitation has been issued
        reviewerId:
          type: integer
          description: the unique identifier of the user who has received the review invitation
        completed:
          type: boolean
          default: false
          description: It's true if the review has benn completed, false otherwise
        reviewDate:
          type: string
          format: date
          description: the date when the review has been completed by the invited user
        rating:
          type: integer
          minimum: 0
          maximum: 10
          description: the rating of the review the user has completed
        review:
          type: string
          maxLength: 1000
          description: a textual description of the review
        film: 
          type: string
          format: uri
          description: URI to access the film related to the review
        delete: 
          type: string
          format: uri
          description: URI to delete the modification request(if not completed).
        self:
          type: string
          format: uri
          description: URI to access the specific review resource
      required:
        - filmId
        - reviewerId
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid